+++ src/app/(pages)/sign-in/page.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import { SubmitHandler, useForm } from 'react-hook-form'
-
 import { SignInRequest } from '@/types/api/auth.types'
-import { useSignInMutation } from 'queries/useSignIn'
+import { useSignInMutation } from 'queries/auth/useSignIn'
+
 
 export default function LoginPage(): JSX.Element {
   const {
