+++ src/app/(pages)/sign-up/page.tsx
@@ -3,8 +3,10 @@
 // import { useRouter } from 'next/navigation'
 import { SubmitHandler, useForm } from 'react-hook-form'
 
+
 import { SignUpRequest } from '@/types/api/auth.types'
-import { useSignUpMutation } from 'queries/useSignUp'
+import { useSignUpMutation } from 'queries/auth/useSignUp'
+
 
 export default function SignUpPage(): JSX.Element {
   // const router = useRouter()
