+++ src/components/common/text/index.tsx
@@ -1,0 +1,4 @@
+import { Highlight } from './Highlight'
+import { Text } from './Text'
+
+export { Text, Highlight }
